+++ platforms/ios/www/cordova_plugins.js
@@ -1,6 +1,35 @@
 cordova.define('cordova/plugin_list', function(require, exports, module) {
 module.exports = [
     {
+        "file": "plugins/org.apache.cordova.console/www/console-via-logger.js",
+        "id": "org.apache.cordova.console.console",
+        "clobbers": [
+            "console"
+        ]
+    },
+    {
+        "file": "plugins/org.apache.cordova.console/www/logger.js",
+        "id": "org.apache.cordova.console.logger",
+        "clobbers": [
+            "cordova.logger"
+        ]
+    },
+    {
+        "file": "plugins/de.appplant.cordova.plugin.badge/www/badge.js",
+        "id": "de.appplant.cordova.plugin.badge.Badge",
+        "clobbers": [
+            "plugin.notification.badge",
+            "cordova.plugins.notification.badge"
+        ]
+    },
+    {
+        "file": "plugins/org.apache.cordova.dialogs/www/notification.js",
+        "id": "org.apache.cordova.dialogs.notification",
+        "merges": [
+            "navigator.notification"
+        ]
+    },
+    {
         "file": "plugins/de.appplant.cordova.plugin.background-mode/www/background-mode.js",
         "id": "de.appplant.cordova.plugin.background-mode.BackgroundMode",
         "clobbers": [
@@ -19,8 +48,11 @@ module.exports = [
 module.exports.metadata = 
 // TOP OF METADATA
 {
-    "de.appplant.cordova.plugin.background-mode": "0.6.0-dev",
-    "org.apache.cordova.device": "0.2.13-dev"
+    "org.apache.cordova.console": "0.2.12",
+    "de.appplant.cordova.plugin.badge": "0.6.1",
+    "org.apache.cordova.dialogs": "0.2.12-dev",
+    "de.appplant.cordova.plugin.background-mode": "0.6.1",
+    "org.apache.cordova.device": "0.2.14-dev"
 }
 // BOTTOM OF METADATA
 });