+++ plugins/de.appplant.cordova.plugin.background-mode/CHANGELOG.md
@@ -1,7 +1,17 @@
 ## ChangeLog
-#### Version 0.6.0 (not yet released)
+#### Version 0.6.1 (14.12.2014)
+- [enhancement:] Set default settings through `setDefaults`.
+- [enhancement:] New method `isEnabled` to receive if mode is enabled.
+- [enhancement:] New method `isActive` to receive if mode is active.
+- [bugfix:] Events caused thread collision.
+
+
+#### Version 0.6.0 (14.12.2014)
 - [feature:] Android support
+- [feature:] Change Android notification through `configure`.
+- [feature:] `onactivate`, `ondeactivate` and `onfailure` callbacks.
 - [___change___:] Disabled by default
+- [enhancement:] Get default settings through `getDefaults`.
 - [enhancement:] iOS does not require user permissions, internet connection and geo location anymore.
 
 #### Version 0.5.0 (13.02.2014)